+++ python/packages/aws-lambda-sdk/tests/test_internal_extension.py
@@ -1,5 +1,4 @@
 from __future__ import annotations
-import pytest
 import os
 
 from typing_extensions import Final