+++ python/packages/sdk/serverless_sdk/lib/warning_captured_event.py
@@ -43,7 +43,7 @@ def create(
     from .. import serverlessSdk
 
     if (
-        origin == "pythonConsole"
+        origin == "pythonLogging"
         or type != "user"
         or serverlessSdk._settings.disable_captured_events_stdout
     ):