+++ python/packages/sdk/tests/lib/test_warning_captured_event.py
@@ -21,7 +21,7 @@ def assert_protobuf_dict(captured_event, tags, fingerprint=None):
         "eventName": "telemetry.warning.generated.v1",
         "timestampUnixNano": to_protobuf_epoch_timestamp(captured_event.timestamp),
         "tags": convert_tags_to_protobuf(captured_event.tags),
-        "customTags": json.dumps(convert_tags_to_protobuf(tags)),
+        "customTags": json.dumps(tags),
         "customFingerprint": fingerprint,
     }
 
@@ -77,7 +77,7 @@ def test_create_warning_captured_event_from_python_console():
     # given
     message = "Warning message"
     tags = {"user.tag": "example"}
-    origin = "pythonConsole"
+    origin = "pythonLogging"
 
     # when
     with mock.patch.object(logger, "warning") as mock_logger: