+++ python/packages/sdk/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## 0.3.4 (2023-03-20)
+
+### Bug Fixes
+
+- Fix behind the scenes logging
+
+## 0.3.3 (2023-03-20)
+
+### Features
+
+- Instrument `.error` & `.warning` methods of Python's native `Logger` class.
+
+## 0.3.2 (2023-03-17)
+
+### Bug Fixes
+
+- Fix custom tags serialization
+
 ## 0.3.1 (2023-03-16)
 
 ### Features